refactor(F204): move plugin handlers to src/plugins/ — out of limb framework
Plugin-specific handler code should not live inside domains/limb/
(generic framework). Move to src/plugins/weixin-mp/ — clearly scoped
as plugin implementation, separate from both the limb framework and
the plugin declarations (plugins/weixin-mp/ at monorepo root).

Three-layer separation:
- domains/limb/          → generic YAML parsing + REST executor + token mgmt
- plugins/weixin-mp/     → declarations (YAML, skills) — no TypeScript
- src/plugins/weixin-mp/ → platform-specific invoke handlers (TypeScript)

[宪宪/claude-opus-4-6🐾]
